% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytobandLevelAlteration.R
\name{CytobandLevelAlteration}
\alias{CytobandLevelAlteration}
\title{CytobandLevelAlteration()}
\usage{
CytobandLevelAlteration(
  Input.Segments,
  Genome = c("hg38", "hg19"),
  Art.len = 0.05
)
}
\arguments{
\item{Input.Segments}{Segment data frame created by MakeSegInput()}

\item{Genome}{Whether samples are aligned to hg38 or hg19}

\item{Art.len}{Threshold for artifact segments (default = 0.05 Mb)}
}
\value{
A tibble containing cytoband level copy number alteration and affected fraction of each cytoband
}
\description{
Function to find cytoband level alteration in each sample using df generated by MakeSegInput()
}
