% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractCnaSignature.R
\name{ExtractCnaSignature}
\alias{ExtractCnaSignature}
\title{ExtractCnaSignature()}
\usage{
ExtractCnaSignature(
  Input.Segments,
  Focal_th = 3,
  Broad_th = 10,
  Art.len = 0.05,
  k_range = 2:10,
  n_run = 100
)
}
\arguments{
\item{Input.Segments}{Segment data frame created by MakeSegInput()}

\item{Focal_th}{Threshold of length for focal alterations (default = 3 Mb)}

\item{Broad_th}{Threshold of length for broad alterations (default = 10 Mb)}

\item{Art.len}{Threshold for artifact segments (default = 0.05 Mb)}

\item{k_range}{K range for NMF}

\item{n_run}{number of runs for NMF}
}
\value{
A list of Signatures with their annotation and contribution to each sample
}
\description{
Function to extract CNA signatures based on df created by MakeSegInput() function.
}
