% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeSegInput.R
\name{MakeSegInput}
\alias{MakeSegInput}
\title{MakeSegInput()}
\usage{
MakeSegInput(
  Segments,
  SampleID_col,
  Chr_col,
  Start_col,
  End_col,
  MajorCN_col,
  MinorCN_col
)
}
\arguments{
\item{Segments}{A data frame of allele-specific segments for all samples.}

\item{SampleID_col}{Name of the column which contains sample IDs.}

\item{Chr_col}{Name of the column which contains chromosome names.}

\item{Start_col}{Name of the column which contains start position of segments.}

\item{End_col}{Name of the column which contains end position of segments.}

\item{MajorCN_col}{Name of the column which contains major copy number of segments.}

\item{MinorCN_col}{Name of the column which contains minor copy number of segments.}
}
\value{
A tibble with standardised columns: SampleID, Chr, Start, End, MinorCN, CopyNumber, len_bp, len_mb.
}
\description{
This function renames user-specified columns to a standard format and computes additional variables useful for copy number analysis.
}
