% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGenomeWideAlteration.R
\name{PlotGenomeWideAlteration}
\alias{PlotGenomeWideAlteration}
\title{PlotGenomeWideAlteration()}
\usage{
PlotGenomeWideAlteration(
  Cytoband.level.data,
  Genome = c("hg38", "hg19"),
  Coverage = 0.8
)
}
\arguments{
\item{Cytoband.level.data}{Data frame of cytoband alterations}

\item{Genome}{Whether samples are aligned to hg38 or hg19}

\item{Coverage}{Threshold for coverage of cytobands needed to be included in the graph (default = 0.8)}
}
\value{
A list of graphs showing the frequency of each alteration in each cytoband across the cohort
}
\description{
Function to graph genome-wide alteration based on cytoband level alteration table generated by CytobandLevelAlteration()
}
