% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSignatures.R
\name{PlotSignatures}
\alias{PlotSignatures}
\title{PlotSignatures()}
\usage{
PlotSignatures(Sig.Results)
}
\arguments{
\item{Sig.Results}{List of results created by ExtractCnaSignature()}
}
\value{
List of bar plots for each identified signatures
}
\description{
Function to generate overview plot of identified signatures
}
